+++ clients/src/main/java/org/apache/kafka/clients/admin/DescribeAclsOptions.java
@@ -34,7 +34,7 @@ public class DescribeAclsOptions extends AbstractOptions<DescribeAclsOptions> {
      *
      */
     // This method is retained to keep binary compatibility with 0.11
-    public DescribeAclsOptions timeoutMs(Integer timeoutMs) {
+    public DescribeAclsOptions timeoutMs(Long timeoutMs) {
         this.timeoutMs = timeoutMs;
         return this;
     }
